percpu: rework memcg accounting
The current implementation of the memcg accounting of the percpu
memory is based on the idea of having a separate set of chunks for
accounted and non-accounted memory. This approach has an advantage
of not wasting any extra memory for memcg data for non-accounted
chunks, however it complicates the code and leads to a less effective
chunk usage due to a lower utilization.

Instead of having two chunk types it's possible to have only one and
allocate the space for memcg data (objcg array) dynamically. It makes
the code simpler in many places (no need to iterate over multiple
chunk types) and saves memory. The only downside is that in a unlikely
event when a percpu allocation succeeds and the allocation of the
objcg array fails, the percpu allocation is passed unaccounted. It's
not a big deal (a single unaccounted allocation will unlikely cause
any systematic problems) and it's how the slab accounting works too.

On my test vm the percpu size just after boot decreased from 7680 kB
to 7488 kB and the number of chunk decreased by 1. It's not a big win,
however on production machines with many chunks the win will be likely
bigger.

Signed-off-by: Roman Gushchin <guro@fb.com>