sched: remove a sched BPF hook in load_balance
The hook, at the start of load_balance, is useless since we now
use kprobe BPF as the alternative.  The reason we use kprobe here
is that we need calltrace, which sched-BPF's context doesn't provide
such parameter for us to properly call the BPF helper which get us
calltrace. I.e. the helper needs context that sched-BPF do not have.